requirements.txt
grpcio
grpcio-tools
transformers
accelerate
